droidboot: add bootloader functionality
We restore threads which monitor for event input and if
there is none for 8 seconds, kexec into the default kernel.
An input event, or fastboot command over USB will cancel
autoboot. This mechanism is off by default.

Aboot.c is broken up and now only has the fastboot command
implementations in it.

'Reboot' is added as an OEM command, with support for the MFLD
OSIP driver. The 'boot' command is heavily dependent on mkbootimg
images and is stubbed out.

Some configuration defines moved to droidboot.h, which can be
overridden in the Android.mk if desired.

Change-Id: I18ea986bb7b25f77091bbef2bd4df92488e453b7
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>