add OSIP writing, reboot command
() "reboot" is a standard Fastboot command. Move it out of
set of oem commands.

() osipX is now a valid flash target. It accepts a stitched
MFLD FSTK image that has exactly one OSII record, and writes
it to OSIP index X.

() Most of the printf() in the UMG update_osip code changed
to use the debug.h API

() Unused code in update_osip.c removed

() Numerous compiler warnings in the UMG code fixed

Change-Id: I873bf311a1af76022ebec34049f73cf11fbf80dd