add droidboot.updatepause option
If this value is nonzero, Droidboot will not automatically apply
a detected SW update in try_update_sw(). Instead it will accept
Fastboot commands as normal, and the semantics of "fastboot continue"
will change to apply the update once invoked.

This is useful for situations where unit-specific provisioning
tasks or tests must be performed before installing the SW.

Change-Id: I3c02a3e5d53b56bacdab3bde6af46147491e51df