support flash command parameterization
() Users can now follow the flash target with a colon and
then a comma-separated list of parameters, described in
more detail in the cmd_flash() comment. Briefly, we have
arguments to support gzip compressed images, destination
device node offsets, and skipping EXT filesystem checks.

() Parsing code added to process these parameters and
put them in a hash table

() The function prototype for flash commands changed so that
the destination parameter is no longer const; we need to modify
it in-place to avoid doing a great many malloc/free operations
when parsing for arguments. There is no need to make it const
anyway as it is never used again by the Fastboot code after it is
passed to the flash function callback.

() Added parameters for file offsets and file append mode to the
named_file_write* functions

() Remove mode bits from named_file_write() open() call since we
set the umask explicitly in main().

Plug-in flash functions can not yet be parameterized; this will
be done in a subsequent patch since it will require a change to
the plug-in UI.

Other than situations which could cause a crash, there is no
sanity checking done on arguments passed in. Garbage in, garbage out.

Change-Id: I5ee4c8cc5faf6c1237ad8cd0c1798eadac6337e1
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>