Fix an intermediate build issue with copying files
Depending on how make schedules parallel jobs, sometimes
the droidboot ramdisk builds before kernel modules are
copied into place. If the ramdisk needs to copy kernel
modules, this will cause the build to fail. This makes
an explicit dependency on kernel for droidboot so to
make sure kernel is built before droidboot.

Change-Id: I84c876d75677a6e944aa2416cbd6202586b59863
Signed-off-by: Daniel Leung <daniel.leung@intel.com>