initial 'oem lock/unlock' implementation
The Fastboot lock/unlock state is tracked via an "OEMLock" EFI
variable. The device is treated as unlocked if and only if the
variable exists and contains the string "0". This value is
persistent even if the entire disk is erased.

Locked devices only support the following fastboot commands:

 oem unlock
 reboot
 reboot-bootloader
 continue
 getvar

Unlocking the device initiates a secure data wipe if the disk
has a valid partition table. Once complete, it enables additional
commands:

 flash
 erase
 format
 boot
 flash:raw
 all other OEM commands

The device may also be re-locked, which initiates an immediate reboot.
The lock state may be queried programmatically via 'getvar locked'.

Right now there is no UI for the unlock operation like on Nexus
devices; 'oem lock' simply erases /data and unlocks the device.
Planned enhancements to the UI will add some confirmation to
this process.

Change-Id: I9a7d16f8a10cbee603d381913bae586d18fd3100
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/175753
Reviewed-by: Lind, Robert <robert.lind@intel.com>