log to kernel instead of stdout
No longer required to declare userfastboot as a console
application (which would refuse to start if the console
was not available)

Ensure debugs all have a trailing newline, it's not
optional anymore.

GPT contents are logged when writing a new GPT.

Start the console at boot if and only if this is a debuggable
system, and don't start it with root privileges.

Change-Id: I311247c54314cf5adb87fee0526dd32ab4fad69b
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/176696
Reviewed-by: Lind, Robert <robert.lind@intel.com>