add bootloader and bootimage sanity checks
We need to do  sanity checks on radio and bootloader images prior
to flashing in unlocked mode. We want to try to prevent the user
from flashing stuff that could render their device
incapable of entering Fastboot mode.

For bootimages, check for the magic value which indicates that it
is an AOSP boot image.

For the bootloader, make sure it is a mountable VFAT partition
and that we can find an expected EFI/BOOT/ file.

Change-Id: I182e9a65a47d4b305d0f96370fc55368159d6e73
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/179278
Reviewed-by: Phillips, Scott D <scott.d.phillips@intel.com>
Reviewed-by: Lind, Robert <robert.lind@intel.com>