respond more clearly to the user when loader is locked
Current implementation simply didn't install command handlers
when the loader is locked, and forced a reboot when locking
the loader. The error messages when trying to flash a locked
device were very counter-intuitive.

Now the reboot is not required, and forbidden commands
unambiguously reflect the locked state of the device.

Change-Id: I2e14ad5bf5c60d7e9c7637a5098008dbdbb008b2
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/179298