add UI for 'fastboot OEM unlock'
If an OEM unlock requires a userdata wipe, display a menu on screen
so that the user can confirm this action, using text based on what
appears on Nexus devices.

Additions to microui code taken from bootable/recovery in ICS (the
source of the rest of the microui code), with a few adjustments;
show_menu and show_text are disjoint states.

Change-Id: Iba001f31747398802d73cd85b719e4222f0200af
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>