REVERTME: skip bootloader sanity check
The sanity check does a loopback mount of the bootloader image over
the wire, and even though it is later unmounted and the file deleted,
space is not freed in the tmpfs. This can leads to flashing of other
images failing since we're not accurately reporting the amount of
buffer we have.

Change-Id: Ifccb4189de05c184638df2437c121e458f223729
Tracked-by: GMIN-716
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/188786
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Gumbel, Matthew K <matthew.k.gumbel@intel.com>
Tested-by: Gumbel, Matthew K <matthew.k.gumbel@intel.com>