display pr_error() and pr_info() messages in caller's terminal
To assist with debugging where there is no serial console or graphics
available, any invocation of pr_error() or pr_info() during the
processing of a Fastboot command is now sent back to the client's
terminal as an informational message.

The severity level of some debugs were adjusted in light of this new
policy.

Change-Id: I41c0d505c6c66ad3344c66fbde365f00ef1045c6
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/189219
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Bergeron, Michael <michael.bergeron@intel.com>