userfastboot: ensure data is flushed before rebooting
When the fastboot client sends a command, it expects an acknowledgment
(typically "OKAY" or "FAIL"). When the command is a reboot request,
there is a risk that the server writes the ack to its internal buffers,
and then reboots before the message has been sent.
To eliminate this problem, the fastboot server already calls the
sync() function to flush all file system buffers. This, however, does
not help when the client and server are connected via TCP, as TCP
buffers are unaffected by sync().
To resolve the issue, the file descriptors are closed before rebooting.

Issue: GMIN-2165
Change-Id: I63acc1f691ede641b8b932c4737972b8cc193dc9
Signed-off-by: Hans Odeberg <hans.odeberg@intel.com>
Reviewed-on: https://android.intel.com/217688
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>
Reviewed-by: Boie, Andrew P <andrew.p.boie@intel.com>
Tested-by: Boie, Andrew P <andrew.p.boie@intel.com>