Add "oemvars" flashable pseudo-partition
GMIN OEM variables are stored as EFI variables.  By default, they are
under the fastboot GUID of 1AC80A82-4F0C-456B-9A99-DEBEB431FCC1.

The syntax supports "#-style" end of line comments

Variable settings are specified as "<var> <val>".  Whitespace around
the variable name is removed, as is trailing whitespace at the end
of the line.  The value can otherwise contain any printable
character and is stored as an 8-bit string in the EFI variable's
value.  Non-printable bytes can be encoded with "%xx" URL-style
notation.

A line of the form:

    GUID = xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx

will change the GUID used for subsequent lines.

Change-Id: Iebfd2687311e1ddb46bd27186aac47bf97c41c5b
Signed-off-by: Andy Ross <andrew.j.ross@intel.com>