GMIN-2871: make sure flash gpt does not create partition in wrong device
The handler for flash gpt checks /sys/block/<device>/removable to see if the
storage device is removable. If it detects multiple non removable device it
chooses the largest one as primary. Unfortunately, SD card in ECS does not
have removable set to 1. so fastboot thinks it is internal. Also if it is
larger than internal storage, GPT table is created in the SD card destroying
all its contents.

To overcome this issue, also check  /dev/block/<device>/device/type to see
if its SD. If it is, skip that device. Ideally, kernel should determine
correctly if the storage device is removable. But this extra check helps to
overcome the issue.

Bug: GMIN-2871
Change-Id: I13dec81723f7e6bf8d1b8370dcec18075110649f
Signed-off-by: Manish Regmi <manish.regmi@intel.com>
Reviewed-on: https://android.intel.com/225361
Reviewed-by: Boie, Andrew P <andrew.p.boie@intel.com>
Tested-by: Boie, Andrew P <andrew.p.boie@intel.com>