implement keystore verification
We use OpenSSL ASN.1 APIs to examine the keystore sent by the
user to ensure that it is valid keystore data before accepting
it. The signture is not checked against the OEM key, just that
the data is well-formed.

Some unused junk in userfastboot.h has been removed, and
-Wextra added to CFLAGS.

Issue: GMINL-69
Change-Id: I069e3678e790b06d9fb599e7990ffc1bb80ad6aa
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/227731