Implement factory reset protection
fastboot oem unlock will now only work if the last byte in the
PersistentDataBlock has been set.  This byte is controlled by the
developer options in the settings application, and is disabled by
default.

Because of this new behavior, this change also modifies the assumption
that the device should be considered locked in the absence of the EFI
variable.  This is to allow a blank device to be flashed.

Change-Id: I06602e4eccbd6ade7cb50a5a13ff5292f5e22572
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>