add 'oem get-hashes'
For verifiying the correctness of device provisioning and OTA updates,
we need a method to check the hashes of the bootloader components,
boot images, and system partition.

The old method in KK did not work in user builds as it required
ADB root access. It also didn't consider the correctness of the
entire system image + verity metadata.

This command, runnable in any device state, properly hashes all
the boot imags, files in the EFI System Partition, and the system
image. The values reported back to the user can then be compared
with the same images in the target-files-package.

Issue: GMIN-1291
Change-Id: Ica4d46e8338a42f3830aa57da92d7bcecb19ef4b
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Reviewed-on: https://android.intel.com/257194
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>
Reviewed-by: Bergeron, Michael <michael.bergeron@intel.com>