fix secure-boot status implementation
UEFI specification 2.4. Section 3.3

The platform firmware is operating in secure boot mode if the value of
the SetupMode variable is 0 and the SecureBoot variable is set to 1. A
platform cannot operate in secure boot mode if the SetupMode variable
is set to 1. The SecureBoot variable should be treated as read-only.

Change-Id: Ie80ffa01ac11de43654dad27625da8eecfbc8baf
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: https://android.intel.com/284572
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Boie, Andrew P <andrew.p.boie@intel.com>
Tested-by: Boie, Andrew P <andrew.p.boie@intel.com>